add proper breakpoints by instruction addr
Add a bp and restore data before repl begins again

Refactor to have eval_command choose to break or not depending on the
command evaluated. This way, each command has their own wait semantic so
the main loop does not need to constantly determine if it needs to wait
and do something depending on the situation.

Have the main loop get the status returned from waiting so it can check
for some common things like WIFEXITED.